Aarchitecture for QuackPipe
Init version for quackpipe architecture
